refactor: organizing go files
